Creating script for DB

+++ DumDum/DumDum/Models/Entities/Troop.cs
@@ -6,8 +6,8 @@
         public string TroopType { get; set; }
         public int Level { get; set; }
         public int HP { get; set; }
-        public int Attack { get; set; }
-        public int Defence { get; set; }
+        public double Attack { get; set; }
+        public double Defence { get; set; }
         public double Speed { get; set; }
         public int StartedAt { get; set; }
         public int FinishedAt { get; set; }
